+++ src/main/java/com/verdantartifice/primalmagick/client/gui/widgets/research_table/ItemTagProjectMaterialWidget.java
@@ -10,9 +10,12 @@ import javax.annotation.Nonnull;
 import com.verdantartifice.primalmagick.client.util.GuiUtils;
 import com.verdantartifice.primalmagick.common.theorycrafting.ItemTagProjectMaterial;
 
+import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.network.chat.Style;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
@@ -58,8 +61,17 @@ public class ItemTagProjectMaterialWidget extends AbstractProjectMaterialWidget<
         Minecraft mc = Minecraft.getInstance();
         ItemStack stack = this.getStackToDisplay();
         List<Component> textList = new ArrayList<>();
-        textList.add(stack.getHoverName().copy().withStyle(stack.getItem().getRarity(stack).getStyleModifier()));
-        stack.getItem().appendHoverText(stack, mc.level, textList, mc.options.advancedItemTooltips ? TooltipFlag.Default.ADVANCED : TooltipFlag.Default.NORMAL);
+        MutableComponent nameComponent = stack.getHoverName().copy();
+        if (nameComponent.getStyle().equals(Style.EMPTY)) {
+            nameComponent = nameComponent.withStyle(stack.getItem().getRarity(stack).getStyleModifier());
+        }
+        if (stack.hasCustomHoverName()) {
+            nameComponent = nameComponent.withStyle(nameComponent.getStyle().applyFormat(ChatFormatting.ITALIC));
+        }
+        textList.add(nameComponent);
+        if (ItemStack.shouldShowInTooltip(stack.getHideFlags(), ItemStack.TooltipPart.ADDITIONAL)) {
+            stack.getItem().appendHoverText(stack, mc.level, textList, mc.options.advancedItemTooltips ? TooltipFlag.Default.ADVANCED : TooltipFlag.Default.NORMAL);
+        }
         return textList;
     }