+++ changelog.md
@@ -5,6 +5,7 @@ v4.0.7
 * Fixed tooltip Z-order issue with Runecarving Table screen
 * Fixed tooltip Z-order issue with Research Table screen
 * Changed research projects to use the Forge milk tag instead of just the milk bucket
+* Added new theorycrafting project to the pool
 
 v4.0.6
 ------